+++ src/main/java/datastreams_knu/bigpicture/stock/agent/StockCrawlingAgent.java
@@ -6,8 +6,8 @@ import datastreams_knu.bigpicture.common.util.WebClientUtil;
 import datastreams_knu.bigpicture.stock.agent.dto.KoreaStockCrawlingDto;
 import datastreams_knu.bigpicture.stock.agent.dto.StockInfoDto;
 import datastreams_knu.bigpicture.stock.agent.dto.USStockCrawlingDto;
-import datastreams_knu.bigpicture.stock.domain.Stock;
-import datastreams_knu.bigpicture.stock.domain.StockInfo;
+import datastreams_knu.bigpicture.stock.entity.Stock;
+import datastreams_knu.bigpicture.stock.entity.StockInfo;
 import datastreams_knu.bigpicture.stock.repository.StockRepository;
 import dev.langchain4j.agent.tool.Tool;
 import lombok.RequiredArgsConstructor;
